Health refresh bug fixed
Fixed a bug in DisplayCharacter where the logic for cards taking damage worked (currentHealth variable was changed), but it failed to refresh in the game view. Fixed by calling the show() method immediately after damage is dealt. Also made a start on creating a JSON file to store character card data.